% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abtest-package.R
\docType{package}
\name{abtest-package}
\alias{abtest}
\alias{abtest-package}
\title{abtest: Bayesian A/B Testing}
\description{
Conduct A/B tests for two conditions and/or two groups. The Shiny App will show the marginal and conditional conversion rates and contrasts.
}
\author{
\strong{Maintainer}: W. Jake Thompson \email{wjakethompson@gmail.com} (\href{https://orcid.org/0000-0001-7339-0300}{ORCID})

}
\keyword{internal}
